finder refactor and unit test

+++ AtomDB/src/main/java/org/g2n/atomdb/sstIO/IOFileChannelReader.java
@@ -63,7 +63,7 @@ public class IOFileChannelReader extends IOReader {
     }
 
     @Override
-    public void position(int positionToMove) throws IOException {
+    public void position(long positionToMove) throws IOException {
         channel.position(positionToMove);
     }
